buy sell box component 화, 스켈레톤 처리
