+++ src/components/BuySellBoxSkeleton.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+export const BuySellBoxSkeleton = () => {
+  return (
+    <div className="s_buysellBox_wrap">
+      <div className="s_spacing_50px_line"></div>
+      <div className="s_skeleton_box"></div>
+      <div className="s_skeleton_box"></div>
+      <div className="s_skeleton_box"></div>
+    </div>
+  );
+};
